+++ Idle_Game/app/src/main/java/com/example/idle_game/util/PermissonRequests.kt
@@ -1,0 +1,39 @@
+package com.example.idle_game.util
+
+import android.Manifest
+import android.app.Activity
+import android.content.pm.PackageManager
+import android.os.Build
+import androidx.core.app.ActivityCompat
+import androidx.core.content.ContextCompat
+
+fun checkAndRequestNotificationPermission(activity: Activity) {
+    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+        if (ContextCompat.checkSelfPermission(
+                activity,
+                android.Manifest.permission.POST_NOTIFICATIONS
+            ) != PackageManager.PERMISSION_GRANTED) {
+            ActivityCompat.requestPermissions(
+                activity,
+                arrayOf(android.Manifest.permission.POST_NOTIFICATIONS),
+                1
+            )
+        }
+    }
+}
+
+fun checkAndRequestBluetoothPermissions(activity: Activity) {
+    val permissions = listOf(
+        Manifest.permission.BLUETOOTH_CONNECT,
+        Manifest.permission.BLUETOOTH_SCAN,
+        Manifest.permission.BLUETOOTH_ADVERTISE
+    )
+
+    val permissionsToRequest = permissions.filter {
+        ContextCompat.checkSelfPermission(activity, it) != PackageManager.PERMISSION_GRANTED
+    }
+
+    if (permissionsToRequest.isNotEmpty()) {
+        ActivityCompat.requestPermissions(activity, permissionsToRequest.toTypedArray(), 1)
+    }
+}
